requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
Django==5.0.7
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
idna==3.7
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
